fix: run the task-join latency verification sum in float64
The first real full-length inc execution (3,555.96 s, completed and
hash-verified) was refused by VEC-04 reconciliation at time index 2824.
Measured cause: the verification-side float32 reduction's own rounding
(12.0 ms) exceeded the tolerance allowance (11.67 ms) at incident
magnitudes — 3,649 active tasks per step, sums near 3.9e7 ms — while the
evidence itself reconciles in float64 on all 3,600 steps with worst
relative error 2.4e-7 against the unchanged 3e-7 tolerance. The
evaluator's published data was never wrong; the verifier's arithmetic
was. One-word dtype correction; tolerances, contracts, and fingerprints
untouched. Adds a wide-trace fixture test at incident magnitudes; the
deterministic revert-guard is the real published inc run behind the
env-gated integration chain test, which now passes 3/3.

+++ src/traffictwin/integration/vec_task_join/service.py
@@ -102,7 +102,10 @@ def build_task_join_report(
     if not np.array_equal((task_met & task_active).sum(axis=(1, 2)), perstep["done"]):
         raise VecTaskJoinError("per-task deadline successes do not match the per-step stream")
     task_latency = np.where(task_active, pertask["task_lat_ms"], np.float32(0.0))
-    latency_by_step = task_latency.sum(axis=(1, 2), dtype=np.float32)
+    # The verification sum runs in float64: a float32 reduction's own rounding at
+    # wide-trace magnitudes (thousands of active tasks per step) exceeds the
+    # tolerance and would reject evidence whose float64 sum reconciles exactly.
+    latency_by_step = task_latency.astype(np.float64).sum(axis=(1, 2))
     for index, (observed, expected) in enumerate(
         zip(latency_by_step, perstep["lat_sum"], strict=True)
     ):
